scheduler/requirements.txt
asyncpg==0.29.0
redis==5.0.4
prometheus-client==0.20.0
python-dotenv==1.0.1
